+++ lib/src/memory_impl.dart
@@ -2,4 +2,6 @@ part of bigcargo;
 
 class MemoryCargo extends MemoryImpl implements Cargo {
   
+  MemoryCargo({String collection: ""}) : super(collection: collection);
+  
 }